src/router/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.88.0
google-genai>=1.14.0
litellm>=1.0.0
python-dotenv>=1.0.0
